+++ CHANGES.rst
@@ -4,6 +4,11 @@ Changelog
 1.1 (unreleased)
 ----------------
 
+- Remove magic path resolving behavior for resources. Path no longer gets
+  overwritten in resolver. It gets now aquired from parent if not set on
+  resource or resource group.
+  [rnix]
+
 - Introduce ``parent`` attribute on ``ResourceMixin``. Gets set in
   ``ResourceGroup.add`` method to provide hierarchy information.
   [rnix]